Added the dynamically linked library for CVC3 on the mac. The libcvc3jni.dylib is actually built for a 64-bit architecture machine. ToDo: figure out a generic build solution so that everybody can use it easily
Also updated the libcvc3.jar since there is a small API change QueryResult is now SatResult. The latest version of cvc3 is used to build.

Also in the eclipse configuration for run-jpf-symbc had to create another launch configuration since the environment variable to link dynamic libraries is DYLD_LIBRARY_PATH on macs while it is LD_LIBRARY_PATH on linux and other systems. So now there is a run-jpf-symbc-mac in the eclipse configurations.